% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualized.R
\name{plotEnhancedProportions}
\alias{plotEnhancedProportions}
\title{Plot Enhanced-Resolution Proportions}
\usage{
plotEnhancedProportions(enhanced_results, cell_type = NULL)
}
\arguments{
\item{enhanced_results}{Results from \code{\link{enhanceResolution}}.}

\item{cell_type}{Cell type to plot. Defaults to the first one.}
}
\value{
A ggplot object.
}
\description{
Plot Enhanced-Resolution Proportions
}
