% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDDGP-package.R
\docType{package}
\name{SDDGP-package}
\alias{SDDGP}
\alias{SDDGP-package}
\title{SDDGP: Spatial Deconvolution via Deep Gaussian Processes}
\description{
A two-stage Bayesian framework for spatial transcriptomics
deconvolution with full posterior uncertainty quantification.
}
\author{
\strong{Maintainer}: Xiao Yang \email{njjyxl1006@163.com}

}
